+++ item-creation/livescripts/item_examples.ts
